+++ docs/.vitepress/theme/composables/friendly.ts
@@ -18,7 +18,7 @@ export function useFriendly() {
 
     // 定义数据源URL列表，按优先级排序
     const urls = [
-      'https://sponsor.wot-design-uni.cn/friendly.json',
+      'https://sponsor.wot-ui.cn/friendly.json',
       'https://wot-sponsors.pages.dev/friendly.json'
     ]
 
@@ -26,7 +26,7 @@ export function useFriendly() {
     const fetchData = async () => {
       for (const url of urls) {
         try {
-          const response = await axios.get(url, {
+          const response = await axios.get(url + '?t=' + Date.now(), {
             timeout: 5000 // 设置5秒超时
           })
           return response.data && response.data.links ? response.data.links : [] // 成功获取数据后直接返回