+++ docs/.vitepress/theme/composables/sponsor.ts
@@ -1,12 +1,3 @@
-/*
- * @Author: weisheng
- * @Date: 2023-08-01 11:12:05
- * @LastEditTime: 2025-05-07 23:07:17
- * @LastEditors: weisheng
- * @Description:
- * @FilePath: /wot-design-uni/docs/.vitepress/theme/composables/sponsor.ts
- * 记得注释
- */
 import { ref, onMounted } from 'vue'
 import axios from 'axios'
 
@@ -21,7 +12,7 @@ export function useSponsor() {
 
     // 定义数据源URL列表，按优先级排序
     const urls = [
-      'https://sponsor.wot-design-uni.cn/wot-design-uni.json',
+      'https://sponsor.wot-ui.cn/wot-design-uni.json',
       'https://wot-sponsors.pages.dev/wot-design-uni.json'
     ]
 
@@ -29,7 +20,7 @@ export function useSponsor() {
     const fetchData = async () => {
       for (const url of urls) {
         try {
-          const response = await axios.get(url, {
+          const response = await axios.get(url + '?t=' + Date.now(), {
             timeout: 5000 // 设置5秒超时
           })
           return response.data // 成功获取数据后直接返回